Crate wasmer_compiler_llvm

Modules

abi 🔒

compiler 🔒

config 🔒

error 🔒

object_file 🔒

trampoline 🔒

translator 🔒

Structs

LLVM

LLVMCompiler

A compiler that compiles a WebAssembly module with LLVM, translating the Wasm to LLVM IR, optimizing it and then translating to assembly.

Enums

CompiledKind

The compiled function kind, used for debugging in the LLVMCallbacks.

LLVMOptLevel

Defines the optimization level used to compile a Module.

Traits

LLVMCallbacks

Callbacks to the different LLVM compilation phases.

Type Aliases

InkwellMemoryBuffer

The InkWell MemoryBuffer type

InkwellModule

The InkWell ModuleInfo type