Module wasmer_compiler_llvm::abi

Modules

aarch64_systemv 🔒

x86_64_systemv 🔒

Traits

Abi

We need to produce different LLVM IR for different platforms. (Contrary to popular knowledge LLVM IR is not intended to be portable in that way.) This trait deals with differences between function signatures on different targets.

Functions

get_abi