Module wasmer_compiler_llvm::config

Structs

LLVM

Enums

CompiledKind
The compiled function kind, used for debugging in the LLVMCallbacks.

LLVMOptLevel
Defines the optimization level used to compile a Module.

Traits

LLVMCallbacks
Callbacks to the different LLVM compilation phases.

Type Definitions

InkwellMemoryBuffer
The InkWell MemoryBuffer type

InkwellModule
The InkWell ModuleInfo type