wasmer_compiler_llvm

Struct LLVM

pub struct LLVM {
    pub(crate) enable_nan_canonicalization: bool,
    pub(crate) enable_g0m0_opt: bool,
    pub(crate) enable_verifier: bool,
    pub(crate) enable_perfmap: bool,
    pub(crate) opt_level: LLVMOptLevel,
    is_pic: bool,
    pub(crate) callbacks: Option<LLVMCallbacks>,
    pub(crate) middlewares: Vec<Arc<dyn ModuleMiddleware>>,
    pub(crate) num_threads: NonZero<usize>,
}

Fields
enable_nan_canonicalization: boolenable_g0m0_opt: boolenable_verifier: boolenable_perfmap: boolopt_level: LLVMOptLevelis_pic: boolcallbacks: Option<LLVMCallbacks>middlewares: Vec<Arc<dyn ModuleMiddleware>>

The middleware chain.

num_threads: NonZero<usize>

Number of threads to use when compiling a module.

Implementations

impl LLVM

pub fn new() -> Self

Creates a new configuration object with the default configuration specified.

pub fn opt_level(&mut self, opt_level: LLVMOptLevel) -> &mut Self

The optimization levels when optimizing the IR.

pub fn enable_pass_params_opt(&mut self) -> &mut Self

(warning: experimental) Pass the value of the first (#0) global and the base pointer of the first (#0) memory as parameter between guest functions.

pub fn num_threads(&mut self, num_threads: NonZero<usize>) -> &mut Self

pub fn callbacks(&mut self, callbacks: Option<LLVMCallbacks>) -> &mut Self

Callbacks that will triggered in the different compilation phases in LLVM.

fn reloc_mode(&self, binary_format: BinaryFormat) -> RelocMode

fn code_model(&self, binary_format: BinaryFormat) -> CodeModel

pub(crate) fn target_operating_system(&self, target: &Target) -> OperatingSystem

pub(crate) fn target_binary_format(&self, target: &Target) -> BinaryFormat

fn target_triple(&self, target: &Target) -> TargetTriple

pub fn target_machine(&self, target: &Target) -> TargetMachine

Generates the target machine for the current target

pub(crate) fn target_machine_with_opt( &self, target: &Target, enable_optimization: bool, ) -> TargetMachine