Struct wasmer_compiler_llvm::LLVM

pub struct LLVM {
    pub(crate) enable_nan_canonicalization: bool,
    pub(crate) enable_verifier: bool,
    pub(crate) opt_level: LLVMOptLevel,
    is_pic: bool,
    pub(crate) callbacks: Option<Arc<dyn LLVMCallbacks>>,
    pub(crate) middlewares: Vec<Arc<dyn ModuleMiddleware>>,
}

Fields
enable_nan_canonicalization: boolenable_verifier: boolopt_level: LLVMOptLevelis_pic: boolcallbacks: Option<Arc<dyn LLVMCallbacks>>middlewares: Vec<Arc<dyn ModuleMiddleware>>

The middleware chain.

Implementations

impl LLVM

pub fn new() -> Self

Creates a new configuration object with the default configuration specified.

pub fn opt_level(&mut self, opt_level: LLVMOptLevel) -> &mut Self

The optimization levels when optimizing the IR.

pub fn callbacks(
&mut self,
callbacks: Option<Arc<dyn LLVMCallbacks>>
) -> &mut Self

Callbacks that will triggered in the different compilation phases in LLVM.

fn reloc_mode(&self) -> RelocMode

fn code_model(&self) -> CodeModel

fn target_triple(&self, target: &Target) -> TargetTriple

pub fn target_machine(&self, target: &Target) -> TargetMachine

Generates the target machine for the current target